+++ lambda-buffers-compiler/src/LambdaBuffers/Compiler/TypeClassCheck/Utils.hs
@@ -1,10 +1,6 @@
 {-# LANGUAGE OverloadedLabels #-}
 
 module LambdaBuffers.Compiler.TypeClassCheck.Utils (
-  -- exports for Validate & Tests
-  type Instance,
-  TypeClassError (..),
-  BasicConditionViolation (..),
   lookupOr,
   checkInstance,
   Tagged (..),
@@ -56,18 +52,25 @@ import LambdaBuffers.Compiler.ProtoCompat.Types qualified as P (
 import LambdaBuffers.Compiler.TypeClassCheck.Pat (Exp, Literal (ModuleName), Pat (AppP, ConsP, DecP, LabelP, LitP, NilP, ProdP, RecP, RefP, SumP, VarP))
 
 import Data.Kind (Type)
-import LambdaBuffers.Compiler.TypeClassCheck.Pretty (pointies, (<///>))
-import LambdaBuffers.Compiler.TypeClassCheck.Solve (Overlap (Overlap))
 import Prettyprinter (
   Pretty (pretty),
-  hcat,
-  indent,
-  line,
   nest,
   prettyList,
-  punctuate,
   vcat,
   (<+>),
+ )
+
+import LambdaBuffers.Compiler.TypeClassCheck.Errors (
+  BasicConditionViolation (OnlyTyVarsInHead, TyConInContext),
+  Instance,
+  TypeClassError (
+    BadInstance,
+    ClassNotFoundInModule,
+    MissingModuleInstances,
+    MissingModuleScope,
+    UnknownClass,
+    UnknownModule
+  ),
  )
 
 -- I believe this is a comonad, but the import isn't worth it here
@@ -96,99 +99,6 @@ instance Show a => Show (Tagged a) where
 instance Pretty a => Pretty (Tagged a) where
   pretty (Tag _ a) = pretty a
 -- *Here* it's useful to distinguish them
-type Instance = Rule Pat
-
-{- Some of these are perfunctory & used to keep functions total.
--}
-data TypeClassError
-  = UnknownClass FQClassName P.SourceInfo -- this might need split into two, investigate further
-  | UnknownModule P.ModuleName -- internal, no sourceInfo (doesn't make sense)
-  | MissingModuleInstances P.ModuleName -- internal, no sourceInfo (doesn't make sense)
-  | MissingModuleScope P.ModuleName -- internal, no sourceinfo (doesn't make sense)
-  | ClassNotFoundInModule Text [Text] -- this one's weird. there's not really one place in the source that triggers it. come back to it later
-  | LocalTyRefNotFound T.Text P.ModuleName P.SourceInfo -- SI is the instance clause that triggered the tyref lookup
-  | SuperclassCycleDetected [[FQClassName]] -- No sourceinfo, it's not very useful due to the nature of cycles
-  | FailedToSolveConstraints P.ModuleName [Constraint Exp] Instance P.SourceInfo -- SI is the instance clause that triggered the subgoal that failed (subgoal itself may not exist anywhere in the source)
-  | MalformedTyDef P.ModuleName Exp P.SourceInfo -- SI is the BODY of the exp
-  | BadInstance BasicConditionViolation P.SourceInfo -- SI is the source of the rule that triggered the violation. This might be weird
-  deriving stock (Show, Eq, Generic)
-
-instance Pretty TypeClassError where
-  pretty = \case
-    UnknownClass cref si ->
-      "Error at" <+> pretty si <+> nest 2 ("Unknown class: " <> pretty cref)
-    UnknownModule mn ->
-      "INTERNAL ERROR: Unknown Module" <+> pretty mn
-    MissingModuleInstances mn ->
-      "INTERNAL ERROR: Missing instance data for module" <+> pretty mn
-    MissingModuleScope mn ->
-      "INTERNAL ERROR: Could not determine TypeClass scope for module" <+> pretty mn
-    ClassNotFoundInModule cn mn ->
-      "Error: Expected to find class"
-        <+> pretty cn
-        <+> "in module"
-        <+> pretty mn
-        <+> "but it isn't there!"
-    LocalTyRefNotFound txt mn _ ->
-      "Error: Expected to find a type definition for a type named"
-        <+> pretty txt
-        <+> "in module"
-        <+> pretty mn
-        <+> "but it isn't there!"
-    SuperclassCycleDetected crs ->
-      "Error: Superclass cycles detected in compiler input:"
-        <+> nest
-          2
-          ( vcat
-              . map (hcat . punctuate " => " . map pretty)
-              $ crs
-          )
-    FailedToSolveConstraints mn cs i _ ->
-      "Error: Could not derive instance:"
-        <+> pointies (pretty i)
-          <> line
-          <> line
-          <> indent 2 "in module"
-        <+> pretty mn
-          <> line
-          <> line
-          <> indent 2 "because the following constraint(s) were not satisfied:"
-          <> line
-          <> line
-          <> indent 2 (vcat $ map pretty cs)
-    MalformedTyDef mn xp _ ->
-      "Error: Encountered malformed type definition:"
-        <> line
-        <> indent 2 (pretty xp)
-        <> line
-        <> indent 2 ("in module" <+> pretty mn)
-    BadInstance bcv _ -> pretty bcv
-
-data BasicConditionViolation
-  = TyConInContext Instance (Constraint Pat)
-  | OnlyTyVarsInHead Instance (Constraint Pat)
-  | OverlapDetected Overlap
-  deriving stock (Show, Eq, Generic)
-
-instance Pretty BasicConditionViolation where
-  pretty = \case
-    TyConInContext inst cst ->
-      "Error: Invalid instance declaration!"
-        <///> indent 2 (pretty inst)
-        <///> "The instance constraint"
-        <///> indent 2 (pretty cst)
-        <///> "Contains a type constructor, but may only contain type variables"
-    OnlyTyVarsInHead inst cst ->
-      "Error: Invalid instance declaration!"
-        <///> indent 2 (pretty inst)
-        <///> "The instance constraint"
-        <///> indent 2 (pretty cst)
-        <///> "only contains type variables, but must contain at least one constructor"
-    OverlapDetected (Overlap cst rules) ->
-      "Error: Overlapping instances detected when trying to solve constraint"
-        <+> pretty cst
-          <> line
-          <> indent 2 (vcat (map pretty rules))
 
 {- NOTE: We need different conditions for MPTCs but these are correct for now
